# document-versioning/internal/api/generated.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.2.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// CreateDocumentRequest defines model for CreateDocumentRequest.
type CreateDocumentRequest struct {
	// Content The JSON content of the document
	Content map[string]interface{} `json:"content"`

	// Name Human-readable name for the document
	Name string `json:"name"`
}

// DocumentResponse defines model for DocumentResponse.
type DocumentResponse struct {
	// Content The JSON content of the document
	Content map[string]interface{} `json:"content"`

	// CreatedAt Document creation timestamp
	CreatedAt time.Time `json:"created_at"`

	// Id Unique document identifier
	Id openapi_types.UUID `json:"id"`

	// Name Document name
	Name string `json:"name"`

	// Version Current version number
	Version int `json:"version"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Error Error message
	Error string `json:"error"`
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Status string `json:"status"`
}

// RevertRequest defines model for RevertRequest.
type RevertRequest struct {
	// Version The version number to revert to
	Version int `json:"version"`
}

// UpdateDocumentRequest defines model for UpdateDocumentRequest.
type UpdateDocumentRequest struct {
	// Content The new JSON content of the document
	Content map[string]interface{} `json:"content"`
}

// VersionInfo defines model for VersionInfo.
type VersionInfo struct {
	// CreatedAt When this version was created
	CreatedAt time.Time `json:"created_at"`

	// Version Version number
	Version int `json:"version"`
}

// VersionListResponse defines model for VersionListResponse.
type VersionListResponse struct {
	// CurrentVersion The current (latest) version number
	CurrentVersion int                `json:"current_version"`
	DocumentId     openapi_types.UUID `json:"document_id"`
	Versions       []VersionInfo      `json:"versions"`
}

// GetDocumentParams defines parameters for GetDocument.
type GetDocumentParams struct {
	// Version Specific version number to retrieve
	Version *int `form:"version,omitempty" json:"version,omitempty"`
}

// CreateDocumentJSONRequestBody defines body for CreateDocument for application/json ContentType.
type CreateDocumentJSONRequestBody = CreateDocumentRequest

// UpdateDocumentJSONRequestBody defines body for UpdateDocument for application/json ContentType.
type UpdateDocumentJSONRequestBody = UpdateDocumentRequest

// RevertDocumentJSONRequestBody defines body for RevertDocument for application/json ContentType.
type RevertDocumentJSONRequestBody = RevertRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Create a new document
	// (POST /api/v1/documents)
	CreateDocument(c *gin.Context)
	// Get a document
	// (GET /api/v1/documents/{id})
	GetDocument(c *gin.Context, id openapi_types.UUID, params GetDocumentParams)
	// Update a document
	// (PUT /api/v1/documents/{id})
	UpdateDocument(c *gin.Context, id openapi_types.UUID)
	// Revert to a specific version
	// (POST /api/v1/documents/{id}/revert)
	RevertDocument(c *gin.Context, id openapi_types.UUID)
	// List document versions
	// (GET /api/v1/documents/{id}/versions)
	ListVersions(c *gin.Context, id openapi_types.UUID)
	// Health check
	// (GET /health)
	HealthCheck(c *gin.Context)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// CreateDocument operation middleware
func (siw *ServerInterfaceWrapper) CreateDocument(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CreateDocument(c)
}

// GetDocument operation middleware
func (siw *ServerInterfaceWrapper) GetDocument(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetDocumentParams

	// ------------- Optional query parameter "version" -------------

	err = runtime.BindQueryParameter("form", true, false, "version", c.Request.URL.Query(), &params.Version)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter version: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetDocument(c, id, params)
}

// UpdateDocument operation middleware
func (siw *ServerInterfaceWrapper) UpdateDocument(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UpdateDocument(c, id)
}

// RevertDocument operation middleware
func (siw *ServerInterfaceWrapper) RevertDocument(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.RevertDocument(c, id)
}

// ListVersions operation middleware
func (siw *ServerInterfaceWrapper) ListVersions(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ListVersions(c, id)
}

// HealthCheck operation middleware
func (siw *ServerInterfaceWrapper) HealthCheck(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.HealthCheck(c)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.POST(options.BaseURL+"/api/v1/documents", wrapper.CreateDocument)
	router.GET(options.BaseURL+"/api/v1/documents/:id", wrapper.GetDocument)
	router.PUT(options.BaseURL+"/api/v1/documents/:id", wrapper.UpdateDocument)
	router.POST(options.BaseURL+"/api/v1/documents/:id/revert", wrapper.RevertDocument)
	router.GET(options.BaseURL+"/api/v1/documents/:id/versions", wrapper.ListVersions)
	router.GET(options.BaseURL+"/health", wrapper.HealthCheck)
}